Modify page ranges for A6A5A1x2 COMBO
Originally, links on A6A5A1x2 COMBO would go only to the left page,
which is nonfunctional.  This commit makes it so that the left page
links to the right page, and the right page links to whatever is the
next page for a particular character.  As a result, we no longer use
the {page}-{page}-2 format for A6A5A1x2 pages, which told the compiler
to skip a page when linking one page to the next.  For example,

  10-14-2

links page 10 to page 12, and page 12 to page 14.

Without using that format, the only thing to keep track of is to make
sure the end of the page range is odd, which corresponds to the right
page.  This is required because the left page is non-functional, so we
should also include that characters links on the right page, where
they can be clicked.