Add all links before styling, style using nth-last-child
This fixes a bug where links on the Terezi retcon pages would be
mislabeled when the retcon link are hidden (for example, when New
Reader Mode is on).

This happens because the mod calculated the CSS based on the link
being present (so the first added link has index 3), and after the
collection removes the retcon link, the first added link should be
index 2, but the CSS is still styling index 3.  This results in the
link either being unstyled or styled as the wrong character.

This commit instead adds all links to the page first and counts how
many links were added, and only after that does it style the links.
It also selects the links using div:nth-last-child(${linkIndex}),
which means if four links were added, the first added link will have
and index of 4, the second an index of 3, the third an index of 2, and
the fourth an index of 1.  The links are then styled going down the
index, starting at 4 and ending at 1.  Because of this change, the
index for the added links remain constant no matter if the retcon page
is there or not.