Add new option: hide original link if a character links to same page
When the original link and a character's link both go to the same
place, this option will hide the original link.  As a side effect, it
also allows keyboard navigation when there is only one link on that
page, which is convenient.  The terezi stuff is because the original
link on those pages is the second link in the array rather than the
first.

This commit also removes an extra "// Add in missing pageNext" section
since that code moved elsewhere.

Also, some indentation fixes.