Collection crashes with blank style; make style "a" instead
When installing the mod on a fresh install of the collection, an error occurs: Styles must define some sort of body!

Due to the somewhat hacky way that Collide and Act 7 are handled, we need to cache the CollideAct7 style rules.  The collection needs the style before it gives POV Cam the chance to compute it, so on the first start, we gave a blank style: "", which will be quickly overwritten for the next launch.  However, this blank style prevents the collection from starting, until the user quits and tries again.

This change just makes it so the placeholder style is "a" instead of "", which the collection is happy with.